User's Manual
This is the user's manual.

This application is neither novel nor groundbreaking; in fact, there are many similar tools available. However, it was specifically designed and coded with educational purposes in mind. The primary goal of this project is to provide a learning platform for understanding secure password generation and software development. This application was developed exclusively for Debian-based distributions, with a particular focus on Pop!_OS, to ensure compatibility and optimal performance within this environment.